requirements.txt
flask==3.0.0
llama-cpp-python==0.2.67
psycopg2==2.9.9
